+++ NEWS.md
@@ -2,6 +2,8 @@
 - Fixed a fatal error when a design unit instantiated multiple times
   inside a generate statement contains a PSL assertion (#1400).
 - Fixed a crash when no command line arguments are given (#1410).
+- Fixed an issue where `library work` would incorrectly report an error
+  when using the `--preserve-case` analysis option (#1413).
 - Several other minor bugs were resolved (#1401).
 
 ## Version 1.19.0 - 2026-01-31